+++ Cortex/Medable.framework/Headers/Medable.h
@@ -95,6 +95,7 @@
 #import "NSString+ThreeDigitsVersion.h"
 #import "NSString+URLExtensions.h"
 #import "NSData+Mime.h"
+#import "NSDate+TimeZoneChange.h"
 
 #endif